Moved the functions in checks to a check_files module in vcstools
